+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.1.2] - 2021-01-30
+
+### Changed
+
+- Changed Message variables from `$message` to variants,
+  such as `$commandMessage` or `$eventMessage`
+
 ## [2.1.1] - 2021-01-28
 
 ### Added
@@ -51,3 +58,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [2.0.0]: https://github.com/novuso/common/compare/1.0.0...2.0.0
 [2.1.0]: https://github.com/novuso/common/compare/2.0.0...2.1.0
 [2.1.1]: https://github.com/novuso/common/compare/2.1.0...2.1.1
+[2.1.2]: https://github.com/novuso/common/compare/2.1.1...2.1.2